RavenDB-20762 Naming and comments
